+++ template-presets/auth/types.ts
@@ -1,0 +1,6 @@
+export type AuthState = {
+  authorized: boolean;
+  setAuthorized: (authorized: boolean) => void;
+  accessToken: string;
+  setAccessToken: (accessToken: string) => void;
+};